requirements.txt
Django>=3.0,<4.0
requests
djangorestframework
gunicorn==20.0.4